+++ src/views/user/list.vue
@@ -109,7 +109,7 @@ export default {
           render: (h, { props: { row }}) => {
             return (
               <div class='table-action'>
-                <el-button type='danger' size='small' onClick={() => this.delete(row.id)}>删除</el-button>
+                <el-button type='danger' size='small' onClick={() => this.delete(row.id)}>删 除</el-button>
               </div>
             )
           }
@@ -131,9 +131,15 @@ export default {
       })
     },
     delete(id) {
-      userApi.deleteUser(id).then(_ => {
-        this.$message1000('删除成功', 'success')
-        this.fetchData()
+      this.$confirm('此操作将删除该用户, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        userApi.deleteUser(id).then(_ => {
+          this.$message1000('删除成功', 'success')
+          this.fetchData()
+        })
       })
     },
     resetForm() {
@@ -144,5 +150,9 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-
+/deep/ .table-action{
+  span{
+    font-size: 14px;
+  }
+}
 </style>
